service/requirements.txt
Django==3.2.16
Djangorestframework==3.14.0
psycopg2
djoser==2.2.0
aiogram
requests